+++ Robust.Shared/Toolshed/TypeParsers/TypeTypeParser.cs
@@ -10,7 +10,6 @@ using System.Threading.Tasks;
 using Robust.Shared.Console;
 using Robust.Shared.ContentPack;
 using Robust.Shared.IoC;
-using Robust.Shared.Log;
 using Robust.Shared.Maths;
 using Robust.Shared.Toolshed.Errors;
 using Robust.Shared.Toolshed.Syntax;
@@ -56,6 +55,7 @@ internal sealed class TypeTypeParser : TypeParser<Type>
     };
 
     private readonly HashSet<string> _ambiguousTypes = new();
+    private CompletionResult? _optionsCache;
 
     public override void PostInject()
     {
@@ -75,6 +75,8 @@ internal sealed class TypeTypeParser : TypeParser<Type>
                 }
             }
         }
+
+        _optionsCache = CompletionResult.FromHintOptions(Types.Select(x => new CompletionOption(x.Key)), "C# level type");
     }
 
     public override bool TryParse(ParserContext parserContext, [NotNullWhen(true)] out object? result, out IConError? error)
@@ -168,8 +170,7 @@ internal sealed class TypeTypeParser : TypeParser<Type>
         string? argName)
     {
         // TODO: Suggest generics.
-        var options = Types.Select(x => new CompletionOption(x.Key));
-        return ValueTask.FromResult<(CompletionResult? result, IConError? error)>((CompletionResult.FromHintOptions(options, "C# level type"), null));
+        return ValueTask.FromResult<(CompletionResult? result, IConError? error)>((_optionsCache, null));
     }
 }
 
@@ -177,9 +178,7 @@ public record struct ExpectedNextType() : IConError
 {
     public FormattedMessage DescribeInner()
     {
-        var msg = new FormattedMessage();
-        msg.AddText($"Expected another type in the generic arguments.");
-        return msg;
+        return FormattedMessage.FromUnformatted("Expected another type in the generic arguments.");
     }
 
     public string? Expression { get; set; }
@@ -191,9 +190,7 @@ public record struct ExpectedGeneric() : IConError
 {
     public FormattedMessage DescribeInner()
     {
-        var msg = new FormattedMessage();
-        msg.AddText($"Expected a generic type, did you forget the angle brackets?");
-        return msg;
+        return FormattedMessage.FromUnformatted("Expected a generic type, did you forget the angle brackets?");
     }
 
     public string? Expression { get; set; }
@@ -205,9 +202,7 @@ public record struct UnknownType(string T) : IConError
 {
     public FormattedMessage DescribeInner()
     {
-        var msg = new FormattedMessage();
-        msg.AddText($"The type {T} is not known and cannot be used.");
-        return msg;
+        return FormattedMessage.FromUnformatted($"The type {T} is not known and cannot be used.");
     }
 
     public string? Expression { get; set; }
@@ -220,9 +215,7 @@ internal record struct TypeIsSandboxViolation(Type T) : IConError
 {
     public FormattedMessage DescribeInner()
     {
-        var msg = new FormattedMessage();
-        msg.AddText($"The type {T.PrettyName()} is not permitted under sandbox rules.");
-        return msg;
+        return FormattedMessage.FromUnformatted($"The type {T.PrettyName()} is not permitted under sandbox rules.");
     }
 
     public string? Expression { get; set; }